requirements.txt
django>=4.2,<5.0
python-docx>=1.1.0
